Proper settings page, no more MAIN_SCHEDULE_ID, CurrentScheduleContext
The Current Schedule is now managed by the API through an `init` endpoint. This opens up the ability to create new schedules through the site itself and toggle between them at any point, which the General Settings page now lets you do.

The goal is to avoid needing to manually touch the database at any point when setting up a new event, and this gets relatively close to that point.