Always query ScheduleStore for selectedEntry in LiveDashboard
When this component managed a full `selectedEntry` in its own internal state, that entry would _not_ get updated when the user updated anything other than the currentEntry. The isolated object meant that even when the Store got updated with new information (like a run's timer starting), the `selectedEntry` wouldn't pick up the changes, so components further down the line that depended on just the prop rather than the Store value wouldn't know to re-render or use new information.